07-Redis_Base_Primaria/requirements.txt
fastapi[all]
uvicorn
gunicorn
pydantic
asyncpg
sqlalchemy
redis_om==0.0.27